Slightly refactor event filters
The user was unable to input e's or k's or use arrow keys in text
boxes. While fixing this bug, which involved having the event filter
allow for further processing of all events, it turned out that a more
granular installation of event filters was needed. Work in progress
for now (but the original bug was fixed).